Biorxiv preprint sim scripts/v2.0 (#99)
* adds files for bioRxiv 281717

* adds network simulation scripts

* work on code documentation for network example

* 2.0rc3->2.0

* removed some commented out code

* adds pip requirements file for the example

* adds pip requirements file for the example

* fix path to pathways files

* must decode to utf-8 on Python3

* unzip must be done with wildcards for morpho types

* mention SLURM job script

* use NEURON 7.5 for TravisCI testing

* prevent h5py v2.8.0 from being used with LFPy

* appears that h5py issue was due to MPI.Reduce with float types, not np.float

* string fix

+++ LFPy/network.py
@@ -1048,10 +1048,10 @@ class Network(object):
                     f1[grp] = np.zeros(shape, dtype=dtype)
                 for key, value in f0[grp].items(): 
                     if RANK == 0:
-                        recvbuf = np.zeros(shape)
+                        recvbuf = np.zeros(shape, dtype=np.float)
                     else:
                         recvbuf = None
-                    COMM.Reduce(value.value, recvbuf, op=op, root=0)
+                    COMM.Reduce(value.value.astype(np.float), recvbuf, op=op, root=0)
                     if RANK == 0:
                         f1[grp][key] = recvbuf
             f0.close()
